Changed the model of my account database, which now can handle real time updates, which will be further extended to other databases
The Database works like this

The main database class, which calls The DAO(data access object). The benefit is that using Annotations like @Dao,etc I don't need to write hard code, it is handled by Room

Dao interacts with database directly and to avoid creating multiple entities of database,i used a singleton class called as ___Client class. This ensures that only one instance of database is created

But to achieve synchronisation of database, I created a repository class, which add another level of abstraction to the database. This helps me to observe changes to database using live data and also provides basic crud operation,

And a single view model class, handles all of these things